==> JAWRA_Files/04_FacilityAnalysis.R <==
###############################################################################################################################################
############################################### Matched Facility Analysis ##################################################################

#Primary Author: Morgan McCarthy, M.S. Biological Systems Engineering, Virginia Tech
#     Updates by Julie Shortridge, BSE

#-----------------------------------------------------------Purpose---------------------------------------------------------------------------#

# The purpose of this script is to compare withdrawals, discharges, and consumption over time for facilities with matched withdrawal
# and discharge records.


##################################################Library Initilization########################################################################
rm(list=ls())

#library(foreign) #allows for easy manipulation of *.dbf file types
#library(rgdal) #extract files from file geodatabases-like in ArcMap
library(dplyr)#data manipulation package that speeds up grouping, summarizing, ordering. 
#library(XML) #downloads and reads XML file types-used to access ECHORest Services to download facilities
#library(RCurl) #downloads and interacts with web-based content
#library(readxl) #reads excel files
#library(jsonlite)
#library(httr)
#library(proj4)
library(data.table)
#library(stringr)
#library(plotly)
#library(shiny)
#library(lubridate)
#library(RColorBrewer)
library(tibble)
#library(tidyr)
#library(Hmisc)
#library(gridExtra)
#library(ggplot2)
library(DescTools)

options(scipen=999) #Disables scientific notation

################################################################Inputs########################################################################

#----------------------------------------------------Cleaned Discharge Data-----------------------------------------------------------------#
# This discharge data was generated by the ECHO_Timeseries.R script and cleaned in the ECHO_QAQC.R Script

#--Load Discharge Timeseries Data--#
setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis")
ECHO_2010_2017<-read.table("ECHO_2010_2017_QAQC.txt", sep="\t", header=T)
## Unique facilities in discharge data
Fac.list<-unique(ECHO_2010_2017$VPDES.Name)

#--Facilities with Matched Withdrawal Facility--#
Matched<-read.csv("Runninglist_Matches.csv")
Matched<-mutate_if(Matched,is.factor,as.character)

#--Remove Municipal Data because it has a noisy relationship--#
Matched<-subset(Matched,!Matched$Sector=="Municipal")

# This function looks at the list of matched facilities and pulls relevant attributes and DMR data for the discharging facilities
matched_discharge<- function(Matched){
  
Match_ECHO_2010_2017<-subset(ECHO_2010_2017,ECHO_2010_2017$Facility.ID%in%gsub("echo_","",Matched$VPDES.Hydrocode),
                             select=c(1,3:12,19,33,43,44,45,46))
                      #select=c(1,3:9,12,19,33,43,44,45,46)) Morgan selected

colnames(Match_ECHO_2010_2017)<-c("OutfallID","VPDES.Facility.ID","VPDES.Name","Fac.Lat","Fac.Long","Outfall_Type","VPDES.Outfall.Long",
                                  "VPDES.Outfall.Lat", "Discharges_MGD_Orig", "Discharges_MGD_ResNA", "Discharges_MGD",
                                  "Date","County","Waterbody","Use.Type","Fuel_Type","Gen_Cap_MW")
    # Note that "Discharges_MGD" column comes from "Resolved_Measured_Effluent_Med" column in ECHO Data
    # Analyses is being performed on QA/QC data where suspect values are replaced with median
    # 15 values are less than zero

Match_ECHO_2010_2017$MatchID<-Matched$MatchID[match(Match_ECHO_2010_2017$VPDES.Facility.ID,gsub("echo_","",Matched$VPDES.Hydrocode))]
Match_ECHO_2010_2017$Date<-as.Date(Match_ECHO_2010_2017$Date,format="%Y-%m-%d")

#ECHO_2010_2017<-subset(ECHO_2010_2017,select=c(1:9,10,12,19,33,40,43,44,45,46,47))
#colnames(ECHO_2010_2017)<-c("OutfallID","Year","VPDES.Facility.ID","VPDES.Name","Fac.Lat","Fac.Long","Outfall_Type",
#                            "Outfall.Long","Outfall.Lat","Measured_Effluent","Discharges_MGD","Date","County",
#                            "Permit_Type","Waterbody","Use.Type","Fuel_Type","Gen_Cap_MW","Mon_Reported")
#ECHO_2010_2017$Date<-as.Date(ECHO_2010_2017$Date,format="%Y-%m-%d")

#assign("ECHO_2010_2017",ECHO_2010_2017,envir=.GlobalEnv)
assign("Match_ECHO_2010_2017",Match_ECHO_2010_2017,envir=.GlobalEnv)
#write.table(ECHO_2010_2017,file="G:/My Drive/ECHO NPDES/USGS_Consumptive_Use_Updated/ECHO_2010_2017_QAQC_statewide.txt",sep="\t")

}
matched_discharge(Matched)

## Look into changed values from QA/QC
summary(Match_ECHO_2010_2017$Discharges_MGD < 0)  #15 negative values
#neg.discharges<-subset(Match_ECHO_2010_2017, Discharges_MGD < 0)
change.ids1<-Match_ECHO_2010_2017$Discharges_MGD != Match_ECHO_2010_2017$Discharges_MGD_Orig  # 137 values changed
change.ids2<-Match_ECHO_2010_2017$Discharges_MGD != Match_ECHO_2010_2017$Discharges_MGD_ResNA  # 134 values changed

#------------------------------------------------------------------------------------------------------------------------------------------#
#-----------------------------------------------Cleaned Withdrawal Data--------------------------------------------------------------------#
# Withdrawal data was obtained through VDEQ's VA Hydro site, which can only be obtained through collaboration.
# Data was analyzed, coordinates were corrected, and water sectors were classified in VWUDS_QAQC.R Script. 

#--Load Withdrawal Timeseries Data--#
load("G:/My Drive/VT/Research/Virginia/USGS_WUDR_ConsumptiveUse/Analysis/FacilityAnalysis/VWUDS_2010_2017.RData")
VWUDS_2010_2017$Year<-as.factor(VWUDS_2010_2017$Year)

# This function looks at the list of matched facilities and pulls relevant attributes and DMR data for the withdrawing facilities
matched_withdrawal<- function(Matched){

#VWUDS_2010_2017<-as.data.table(VWUDS_2010_2017)
VWUDS_2010_2017<-VWUDS_2010_2017[order(VWUDS_2010_2017[,1],VWUDS_2010_2017[,8],decreasing=F),]
colnames(VWUDS_2010_2017)[3]<-"VWUDS.Facility.ID"
colnames(VWUDS_2010_2017)[10]<-"Old.Use.Type"
colnames(VWUDS_2010_2017)[20]<-"Use.Type"

#--Facilities with Matched Discharge Facility--#
Match_VWUDS_2010_2017<-subset(VWUDS_2010_2017,VWUDS_2010_2017$VWUDS.Facility.ID%in%Matched$VWUDS.HydroID)
Match_VWUDS_2010_2017<-Match_VWUDS_2010_2017%>%
  dplyr::group_by(VWUDS.Facility.ID,Date)%>%
  dplyr::summarise(VWUDS.Name=first(Facility),Withdrawal_mgd=sum(Withdrawals_MGD,na.rm=T),VWUDS.Lat=mean(Corrected_Latitude,na.rm=T),VWUDS.Long=mean(Corrected_Longitude,na.rm=T),
                   VWUDS.Use.Type=first(Use.Type))

colnames(Match_VWUDS_2010_2017)<-c("VWUDS.Facility.ID","Date","VWUDS.Name","Withdrawals_MGD","VWUDS.Lat",
                                   "VWUDS.Long","VWUDS.Use.Type")
Match_VWUDS_2010_2017$MatchID<-Matched$MatchID[match(Match_VWUDS_2010_2017$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]

#assign("VWUDS_2010_2017",VWUDS_2010_2017,envir=.GlobalEnv)
assign("Match_VWUDS_2010_2017",Match_VWUDS_2010_2017,envir=.GlobalEnv)

#save(VWUDS_2010_2017,file="G:/My Drive/ECHO NPDES/USGS_Consumptive_Use_Updated/Code/R Workspaces/VWUDS_2010_2017_statewide.RData")

}
matched_withdrawal(Matched)

#-----------------------------Matched Data---------------------------------#
# Look for "fully" matched results for each month. This function merges DMR entries reported by both the withdrawing and discharging 
# facilities. This ensures that we aren't looking at incomplete data. i.e. facilities that report a discharge but not a withdrawal for that reporting period.

# We are also concerned with energy facilities that report a 0.0 MGD discharge or withdrawal. Is it becuase they didn't report, or because it was actually 0?

fully_matched<- function(Match_VWUDS_2010_2017,Match_ECHO_2010_2017){
full_match_2010_2017<-merge(Match_VWUDS_2010_2017,Match_ECHO_2010_2017,by=c("MatchID","Date"))
full_match_2010_2017<-full_match_2010_2017[,c("VPDES.Facility.ID","OutfallID","VWUDS.Facility.ID","VWUDS.Name","VPDES.Name","Use.Type",
                      "Fac.Lat","Fac.Long","VWUDS.Lat","VWUDS.Long","Date","Withdrawals_MGD","Discharges_MGD","County","Waterbody","Fuel_Type",
                      "Gen_Cap_MW")]


#---------------------Dissolve Flow from multiple outfalls into a single facility value----------#
full_match_2010_2017<-full_match_2010_2017%>%dplyr::group_by(VWUDS.Facility.ID,VPDES.Facility.ID,
                                       VWUDS.Name,VPDES.Name,Use.Type,Fac.Lat,Fac.Long,
                                       VWUDS.Lat,VWUDS.Long,Date,Withdrawals_MGD,
                                       County,Waterbody,Fuel_Type,Gen_Cap_MW)%>%dplyr::summarise(Discharges_MGD=sum(Discharges_MGD,na.rm=T),n_outfalls=n_distinct(OutfallID))
full_match_2010_2017<-full_match_2010_2017[c("VWUDS.Facility.ID","VPDES.Facility.ID","n_outfalls",
                                             "VWUDS.Name","VPDES.Name","Use.Type","Fac.Lat","Fac.Long",
                                             "VWUDS.Lat","VWUDS.Long","Date","Withdrawals_MGD", "Discharges_MGD",
                                             "County","Waterbody","Fuel_Type",
                                             "Gen_Cap_MW")]

full_match_2010_2017<-full_match_2010_2017%>%add_column(CU=(full_match_2010_2017$Withdrawals_MGD-full_match_2010_2017$Discharges_MGD)/full_match_2010_2017$Withdrawals_MGD, .after="Discharges_MGD")


full_match_2010_2017_summary<-full_match_2010_2017%>%dplyr::group_by(VPDES.Facility.ID)%>%
  dplyr::summarise(VWUDS.Facility.ID=first(VWUDS.Facility.ID),n_outfalls=first(n_outfalls),VWUDS.Name=first(VWUDS.Name),VPDES.Name=first(VPDES.Name),
                   Use.Type=first(Use.Type), VPDES.Fac.Lat=first(Fac.Lat),VPDES.Fac.Long=first(Fac.Long),VWUDS.Fac.Lat=first(VWUDS.Lat),VWUDS.Fac.Long=first(VWUDS.Long),
                   Ave_Withdrawal_mgd=mean(Withdrawals_MGD,na.rm=T),Ave_Discharge_mgd=mean(Discharges_MGD,na.rm=T),County=first(County),Waterbody=first(Waterbody),
                   Fuel_Type=first(Fuel_Type),Gen_Cap_MW=first(Gen_Cap_MW))

full_match_2010_2017_summary$Ave_CU<-(full_match_2010_2017_summary$Ave_Withdrawal_mgd-full_match_2010_2017_summary$Ave_Discharge_mgd)/full_match_2010_2017_summary$Ave_Withdrawal_mgd

full_match_2010_2017_summary$VPDES_perc<-Matched$VPDES...total[match(full_match_2010_2017_summary$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]
full_match_2010_2017_summary$VWUDS_perc<-Matched$VWUDS...total[match(full_match_2010_2017_summary$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]

                                 
# Not all facilities report 12 months out of the year. Therefore, if we want to aggregate monthly withdrawals into yearly estimates for each
# facility, we must divide by the number of reporting months.
W_Months<-
  full_match_2010_2017%>%
  dplyr::group_by(VWUDS.Facility.ID,Year=substring(Date,1,4))%>%
  dplyr::count(Month=substring(Date,6,7))%>%
  dplyr::summarise(Mon_Reported=sum(n))

full_match_2010_2017<-full_match_2010_2017%>%add_column(Year=substring(full_match_2010_2017$Date,1,4), .after="Date")
full_match_2010_2017<-merge(full_match_2010_2017,W_Months,by=c("VWUDS.Facility.ID","Year"),all.x=T)



# Saves into global environment
assign("full_match_2010_2017",full_match_2010_2017,envir=.GlobalEnv)
assign("full_match_2010_2017_summary",full_match_2010_2017_summary,envir=.GlobalEnv)

#write.table(full_match_2010_2017,file="G:/My Drive/ECHO NPDES/USGS_Consumptive_Use_Updated/full_match_2010_2017.txt",sep="\t")
#write.table(full_match_2010_2017_summary,file="G:/My Drive/USGS_ConsumptiveUse/Spring, 2019/Statewide Analysis/fully_matched_summary.txt",row.names=F,sep="|")

}
fully_matched(Match_VWUDS_2010_2017,Match_ECHO_2010_2017)

length(unique(full_match_2010_2017$VWUDS.Name))
length(unique(full_match_2010_2017$VPDES.Name))

# How many facilities were excluded becasues they didn't have fully matched withdrawal and discharge entries?
missing<-subset(Matched,! gsub('echo_',"",Matched$VPDES.Hydrocode)%in%full_match_2010_2017_summary$VPDES.Facility.ID)

## Add columns to full_match_2010_2017_summary
##  N.record.vpdes, N.zero.vpdes, Mean.all.vpdes, Mean.zerorm.vpdes
##  N.record.vwuds, N.zero.vwuds, Mean.all.vwuds, Mean.zerorm.vwuds

summary(full_match_2010_2017_summary$Ave_CU)
n.fac<-dim(full_match_2010_2017_summary)[1]
fill<-rep(-999,n.fac)

match.summary<-data.frame(VWUDS.ID=full_match_2010_2017_summary$VWUDS.Facility.ID,
                          VPDES.ID=full_match_2010_2017_summary$VPDES.Facility.ID,
                          VWUDS.Name=full_match_2010_2017_summary$VWUDS.Name,
                          VPDES.Name=full_match_2010_2017_summary$VPDES.Name,
                          Use.Type=full_match_2010_2017_summary$Use.Type,
                          VWUDS.Nrecord=fill, VWUDS.Nzero=fill, VWUDS.mean.all=fill, VWUDS.mean.narm=fill, VWUDS.mean.full=fill, VWUDS.mean.scale=fill,
                          VPDES.Nrecord=fill, VPDES.Nzero=fill, VPDES.mean.all=fill, VPDES.mean.narm=fill, VPDES.mean.full=fill, VPDES.mean.scale=fill)

for (f in 1:n.fac){
  vpdes.id<-full_match_2010_2017_summary$VPDES.Facility.ID[f]
  facility.records<-subset(full_match_2010_2017, VPDES.Facility.ID==vpdes.id)
  match.summary$VWUDS.Nrecord[f] <- dim(facility.records)[1] -  sum(facility.records$Withdrawals_MGD == 0) - sum(is.na(facility.records$Withdrawals_MGD))
  match.summary$VPDES.Nrecord[f] <- dim(facility.records)[1] -  sum(facility.records$Discharges_MGD == 0) - sum(is.na(facility.records$Discharges_MGD))
  match.summary$VWUDS.Nzero[f] <- sum(facility.records$Withdrawals_MGD == 0)
  match.summary$VPDES.Nzero[f] <- sum(facility.records$Discharges_MGD == 0)
  ## identify records with both discharge and withdrawal reported
  wd.records<-!facility.records$Withdrawals_MGD == 0 & !is.na(facility.records$Withdrawals_MGD)
  dc.records<-!facility.records$Discharges_MGD == 0 & !is.na(facility.records$Discharges_MGD)
  full.records<-wd.records==TRUE & dc.records == TRUE  
  ## identify records with both discharge and withdrawal within 2 orders of magnitude of each other
  ratio<-facility.records$Discharges_MGD/facility.records$Withdrawals_MGD
  same.scale<-ratio < 100 & ratio > 0.01
  ## Calculate averages across all records and only months with both wd and discharge reported
  match.summary$VWUDS.mean.all[f] <- mean(facility.records$Withdrawals_MGD,na.rm=TRUE)
  match.summary$VWUDS.mean.narm[f] <- mean(facility.records$Withdrawals_MGD[!facility.records$Withdrawals_MGD == 0],na.rm=TRUE)
  match.summary$VWUDS.mean.full[f] <- mean(facility.records$Withdrawals_MGD[full.records==TRUE],na.rm=TRUE)
  match.summary$VWUDS.mean.scale[f] <- mean(facility.records$Withdrawals_MGD[same.scale==TRUE],na.rm=TRUE)
  match.summary$VPDES.mean.all[f] <- mean(facility.records$Discharges_MGD,na.rm=TRUE)
  match.summary$VPDES.mean.narm[f] <- mean(facility.records$Discharges_MGD[!facility.records$Discharges_MGD == 0],na.rm=TRUE)
  match.summary$VPDES.mean.full[f] <- mean(facility.records$Discharges_MGD[full.records==TRUE],na.rm=TRUE)
  match.summary$VPDES.mean.scale[f] <- mean(facility.records$Discharges_MGD[same.scale==TRUE],na.rm=TRUE)
}

## Calculate consumptive use three ways: all records, nonzero records, and fully matched records only
match.summary$CU_all<-(match.summary$VWUDS.mean.all-match.summary$VPDES.mean.all)/match.summary$VWUDS.mean.all
match.summary$CU_nonzero<-(match.summary$VWUDS.mean.narm-match.summary$VPDES.mean.narm)/match.summary$VWUDS.mean.narm
match.summary$CU_full<-(match.summary$VWUDS.mean.full-match.summary$VPDES.mean.full)/match.summary$VWUDS.mean.full
match.summary$CU_scale<-(match.summary$VWUDS.mean.scale-match.summary$VPDES.mean.scale)/match.summary$VWUDS.mean.scale
match.summary$CV_scale<-match.summary$VWUDS.mean.scale-match.summary$VPDES.mean.scale
  # consumptive volume in MGD
match.summary$CV_scale <- match.summary$CV_scale*0.003785*1000
  # convert MGD to CM per day x 1000

write.csv(match.summary,"MatchSummary.csv",row.names=FALSE)

summary(match.summary$CU_all)     # 3 NA's
summary(match.summary$CU_nonzero) # 13 NA's - removes facilities/months with zero withdrawal
summary(match.summary$CU_full)    # 14 NA's - removes facilities/months with zero withdrawal or discharge
summary(match.summary$CU_scale)   # 17 NA's - removes facilities/months without similarly scaled withdrawal and discharge
summary(match.summary$CV_scale)   # consumptive volume - ranges from -146 to +79 mgd


### Create histograms from CU scale column
# Consumptive use calculated using only facilities/months where:
# both discharge and withdrawal reported 
# both discharge and withdrawal are nonzero
# reported withdrawal and discharge are within two orders of magnitude of each other

hist(match.summary$CU_scale[match.summary$Use.Type=="Energy"], main="Energy") # 12 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Industrial"], main="Industrial") # 46 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Commercial"], main="Commercial") # 48 facilities

sum(match.summary$Use.Type=="Commercial")
## Subset to values between -1 and 1

trim.records<-subset(match.summary,CU_scale > -1) # now only 88 facilities - we lose 1 energy, 12 industrail, and 14 commercial
hist(trim.records$CU_scale[match.summary$Use.Type=="Energy"], main="Energy") # 12 facilities
hist(trim.records$CU_scale[match.summary$Use.Type=="Industrial"], main="Industrial") # 34 facilities
hist(trim.records$CU_scale[match.summary$Use.Type=="Commercial"], main="Commercial") # 34 facilities

sum(trim.records$Use.Type=="Energy")

## Create panel figure - left is all energy, industrial and commercial, right is trimmed                

ppi<-300
setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis")

png(file="Histograms.png",width=6*ppi,height=12*ppi,res=ppi)		
  par(mfcol=c(4,2))
  par(mar=c(3,4,2,1.5))
  ## All records
  hist(match.summary$CU_scale[match.summary$Use.Type=="Energy"], main="Energy (All)", xlab="") #13 facilities
  hist(match.summary$CU_scale[match.summary$Use.Type=="Industrial"], main="Industrial (All)", xlab="") # 46 facilities
  hist(match.summary$CU_scale[match.summary$Use.Type=="Commercial"], main="Commercial (All)", xlab="") # 48 facilities
  hist(match.summary$CU_scale[match.summary$Use.Type=="Aquaculture"], main="Aquaculture (All)", xlab="") # 8 facilities
  ## Trimmed records (between -1 and 1)
  par(mar=c(3,3,2,1.5))
  hist(trim.records$CV_scale[trim.records$Use.Type=="Energy"], main="Energy (Subset)", xlab="", ylab="") #12 facilities
  hist(trim.records$CV_scale[trim.records$Use.Type=="Industrial"], main="Industrial (Subset)", xlab="", ylab="") # 34 facilities
  hist(trim.records$CV_scale[trim.records$Use.Type=="Commercial"], main="Commercial (Subset)", xlab="", ylab="") # 34 facilities
  hist(trim.records$CV_scale[trim.records$Use.Type=="Aquaculture" & trim.records$CU_scale < 0.5], 
      main="Aquaculture (Subset)", xlab="", ylab="") # 8 facilities
dev.off()## Create panel figure - left is all energy, industrial and commercial, right is trimmed                


## Revised Histogram (no trimmed records)
ppi<-300
setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis")

png(file="Histograms_R2.png",width=6*ppi,height=6*ppi,res=ppi)		
par(mfcol=c(2,2))
par(mgp=c(2.5,1,0))
par(mar=c(4,4,2,1.5))
## All records
hist(match.summary$CU_scale[match.summary$Use.Type=="Energy"], main="Thermoelectric", xlab="Consumptive Use") #13 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Industrial"], main="Industrial", xlab="Consumptive Use") # 46 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Commercial"], main="Commercial", xlab="Consumptive Use") # 48 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Aquaculture"], main="Aquaculture", xlab="Consumptive Use") # 8 facilities

dev.off()

## Revised Histogram (left is consumption as fraction of withdrawal, right is consumptive volume)
ppi<-300
setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis")

png(file="Histograms_R3.png",width=6*ppi,height=12*ppi,res=ppi)		
par(mfcol=c(4,2))
par(mar=c(4,4,2,1.5))
## Consumptive fraction (W-D)/W
hist(match.summary$CU_scale[match.summary$Use.Type=="Energy"], main="Thermoelectric", xlab="(W-D)/W") #13 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Industrial"], main="Industrial", xlab="(W-D)/W") # 46 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Commercial"], main="Commercial", xlab="(W-D)/W") # 48 facilities
hist(match.summary$CU_scale[match.summary$Use.Type=="Aquaculture"], main="Aquaculture", xlab="(W-D)/W") # 8 facilities
## Consumptive volume (W-D)
par(mar=c(4,4,2,1.5))
hist(match.summary$CV_scale[match.summary$Use.Type=="Energy"], main="Thermoelectric", xlab="W-D (1000 m^3/day)", ylab="") #12 facilities
hist(match.summary$CV_scale[match.summary$Use.Type=="Industrial"], main="Industrial", xlab="W-D (1000 m^3/day)", ylab="") # 34 facilities
hist(match.summary$CV_scale[match.summary$Use.Type=="Commercial"], main="Commercial", xlab="W-D (1000 m^3/day)", ylab="") # 34 facilities
hist(match.summary$CV_scale[match.summary$Use.Type=="Aquaculture"],main="Aquaculture", xlab="W-D (1000 m^3/day)", ylab="") # 8 facilities
dev.off()


## Summary of matches - number of facilities and volumes
match.count<-data.frame(Use.Type=c("Energy","Industrial","Commercial","Aquaculture"),
                        n.match=rep(NA,4), n.match.scale=rep(NA,4), 
                        withdrawal=rep(NA,4), withdrawal.scale=rep(NA,4),
                        discharge=rep(NA,4), discharge.scale=rep(NA,4))
for (s in 1:4){
  sector.data<-subset(match.summary, Use.Type == paste(match.count$Use.Type[s]))
  match.count$n.match[s]<-dim(sector.data)[1]
  match.count$withdrawal[s]<-sum(sector.data$VWUDS.mean.full,na.rm=TRUE)
  match.count$discharge[s]<-sum(sector.data$VPDES.mean.full,na.rm=TRUE)
  sector.trim<-subset(trim.records, Use.Type == paste(match.count$Use.Type[s]))
  match.count$n.match.scale[s]<-dim(sector.trim)[1]
  match.count$withdrawal.scale[s]<-sum(sector.trim$VWUDS.mean.full,na.rm=TRUE)
  match.count$discharge.scale[s]<-sum(sector.trim$VPDES.mean.full,na.rm=TRUE)
}


##################################################################
## Calculate seasonal summaries
##################################################################
## Use consumptive estimates only for months where 
##  - both discharge and withdrawal are reported and 
##  - within 2 orders of magnitude of each other

## Add columns to match.summary
match.summary$CU_scale_warm<-rep(-999, dim(match.summary)[1])
match.summary$CU_scale_cool<-rep(-999, dim(match.summary)[1])
match.summary$CV_scale_warm<-rep(-999, dim(match.summary)[1])
match.summary$CV_scale_cool<-rep(-999, dim(match.summary)[1])

for (f in 1:n.fac){
  vpdes.id<-full_match_2010_2017_summary$VPDES.Facility.ID[f]
  facility.records<-subset(full_match_2010_2017, VPDES.Facility.ID==vpdes.id)
  ## identify records with both discharge and withdrawal within 2 orders of magnitude of each other
  ratio<-facility.records$Discharges_MGD/facility.records$Withdrawals_MGD
  facility.records$same.scale<-ratio < 100 & ratio > 0.01
  
  # Split by warm/cool season
  months<-as.numeric(format(facility.records$Date, "%m") ) # extract months
  facility.warm<-facility.records[months > 3 & months < 11 ,]
  facility.cool<-facility.records[months < 4 | months > 10 ,]
  
  warm.wd <- sum(facility.warm$Withdrawals_MGD[facility.warm$same.scale==TRUE],na.rm=TRUE)
  warm.d <- sum(facility.warm$Discharges_MGD[facility.warm$same.scale==TRUE],na.rm=TRUE)
  match.summary$CU_scale_warm[f] <- (warm.wd - warm.d) / warm.wd
  match.summary$CV_scale_warm[f] <- (warm.wd - warm.d) *0.003785*1000
  
  cool.wd <- sum(facility.cool$Withdrawals_MGD[facility.cool$same.scale==TRUE],na.rm=TRUE)
  cool.d <- sum(facility.cool$Discharges_MGD[facility.cool$same.scale==TRUE],na.rm=TRUE)
  match.summary$CU_scale_cool[f] <- (cool.wd - cool.d) / cool.wd
  match.summary$CV_scale_cool[f] <- (cool.wd - cool.d) *0.003785*1000
  
}

## Summary Table with number of negative facilities in each sector
fac.seasons <- data.frame(Sector=sector.IDs, n.match=rep(NA,4), n.neg.annual=rep(NA,4),
                          n.neg.warm=rep(NA,4), n.neg.cool=rep(NA,4))
for (s in 1:4){
  s.data <- match.summary[match.summary$Use.Type== sector.IDs[s] ,]
  fac.seasons$n.match[s] <- sum(!is.na(match.summary$CU_scale[match.summary$Use.Type == sector.IDs[s] ]))
  fac.seasons$n.neg.annual[s] <- sum(match.summary$CU_scale[match.summary$Use.Type == sector.IDs[s] ] < 0, na.rm=TRUE)
  fac.seasons$n.neg.warm[s] <- sum(match.summary$CU_scale_warm[match.summary$Use.Type == sector.IDs[s] ] < 0, na.rm=TRUE)
  fac.seasons$n.neg.cool[s] <- sum(match.summary$CU_scale_cool[match.summary$Use.Type == sector.IDs[s] ] < 0, na.rm=TRUE)
}
write.csv(fac.seasons, file="facility_seasons.csv")
#####################################################################################################################
## Extract data for summary table (median and interquartile range for annual, warm and cool-season consumption)
#####################################################################################################################

summary(match.summary$Use.Type)

xSecSummary<-data.frame(All.Nonenergy=rep(NA,4),Energy=rep(NA,4),Industrial=rep(NA,4),Commercial=rep(NA,4),Aquaculture=rep(NA,4),Municipal=rep(NA,4))
row.names(xSecSummary)<- c("Annual", "Warm Season", "Cool Season","P-value")
sector.IDs<-c("Energy", "Industrial", "Commercial", "Aquaculture")
for (s in 1:length(sector.IDs)){
  sector.summary<-subset(match.summary, Use.Type==sector.IDs[s])
  an.values<-quantile(sector.summary$CU_scale, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  warm.values<-quantile(sector.summary$CU_scale_warm, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  cool.values<-quantile(sector.summary$CU_scale_cool, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  xSecSummary[1,s+1]<-paste(round(an.values[2],2), " (", round(an.values[1],2), " - ", round(an.values[3],2), ")", sep="")
  xSecSummary[2,s+1]<-paste(round(warm.values[2],2), " (", round(warm.values[1],2), " - ", round(warm.values[3],2), ")", sep="")
  xSecSummary[3,s+1]<-paste(round(cool.values[2],2), " (", round(cool.values[1],2), " - ", round(cool.values[3],2), ")", sep="")
  # Wilcox rank sum test
  test.table<-cbind (sector.summary$CU_scale_warm, sector.summary$CU_scale_cool)
  test.table<-test.table[!(is.na(test.table[,1]) | is.na(test.table[,2]) ) ,]
  wt<-wilcox.test(test.table[,1], test.table[,2], alternative = "two.sided", paired=TRUE)
  xSecSummary[4,s+1]<- wt$p.value
}

## All non-energy facilities
nonenergy.summary<-subset(match.summary, Use.Type != "Energy")
an.values<-quantile(nonenergy.summary$CU_scale, probs=c(0.25,0.5,0.75), na.rm=TRUE)
warm.values<-quantile(nonenergy.summary$CU_scale_warm, probs=c(0.25,0.5,0.75), na.rm=TRUE)
cool.values<-quantile(nonenergy.summary$CU_scale_cool, probs=c(0.25,0.5,0.75), na.rm=TRUE)
xSecSummary[1,1]<-paste(round(an.values[2],2), " (", round(an.values[1],2), " - ", round(an.values[3],2), ")", sep="")
xSecSummary[2,1]<-paste(round(warm.values[2],2), " (", round(warm.values[1],2), " - ", round(warm.values[3],2), ")", sep="")
xSecSummary[3,1]<-paste(round(cool.values[2],2), " (", round(cool.values[1],2), " - ", round(cool.values[3],2), ")", sep="")
# Wilcox rank sum test
test.table<-cbind (nonenergy.summary$CU_scale_warm, nonenergy.summary$CU_scale_cool)
test.table<-test.table[!(is.na(test.table[,1]) | is.na(test.table[,2]) ) ,]
wt<-wilcox.test(test.table[,1], test.table[,2], alternative = "two.sided", paired=TRUE)
xSecSummary[4,1]<- wt$p.value

## Revise cross-sectional summary
## Replace quantile values with confidence intervals around median
XSecRevised <- data.frame(All.Nonenergy=rep(NA,4),Energy=rep(NA,4),Industrial=rep(NA,4),Commercial=rep(NA,4),Aquaculture=rep(NA,4),Municipal=rep(NA,4))
row.names(xSecSummary)<- c("Annual", "Warm Season", "Cool Season","P-value")
sector.IDs<-c("Energy", "Industrial", "Commercial", "Aquaculture")
for (s in 1:length(sector.IDs)){
  sector.summary<-subset(match.summary, Use.Type==sector.IDs[s])
  an.CI<-MedianCI(sector.summary$CU_scale,conf.level=0.90, na.rm=TRUE)
  warm.CI<-MedianCI(sector.summary$CU_scale_warm,conf.level=0.90, na.rm=TRUE)
  cool.CI<-MedianCI(sector.summary$CU_scale_cool,conf.level=0.90, na.rm=TRUE)
  XSecRevised[1,s+1]<-paste(round(an.CI[1],2), " (", round(an.CI[2],2), " - ", round(an.CI[3],2), ")", sep="")
  XSecRevised[2,s+1]<-paste(round(warm.CI[1],2), " (", round(warm.CI[2],2), " - ", round(warm.CI[3],2), ")", sep="")
  XSecRevised[3,s+1]<-paste(round(cool.CI[1],2), " (", round(cool.CI[2],2), " - ", round(cool.CI[3],2), ")", sep="")
  # Wilcox rank sum test
  test.table<-cbind (sector.summary$CU_scale_warm, sector.summary$CU_scale_cool)
  test.table<-test.table[!(is.na(test.table[,1]) | is.na(test.table[,2]) ) ,]
  wt<-wilcox.test(test.table[,1], test.table[,2], alternative = "two.sided", paired=TRUE)
  XSecRevised[4,s+1]<- wt$p.value
}

quant.summary <- data.frame(Energy_25=rep(NA,3), Energy_50=rep(NA,3), Energy_75=rep(NA,3),
                            Industrial_25=rep(NA,3),Industial_50=rep(NA,3),Industrial_75=rep(NA,3),
                            Commercial_25=rep(NA,3),Commercial_50=rep(NA,3), Commercial_75=rep(NA,3),
                            Aquaculture_25=rep(NA,3),Aquaculture_50=rep(NA,3), Aquaculture_75=rep(NA,3))
row.names(quant.summary)<- c("Annual", "Warm Season", "Cool Season")
for (s in 1:length(sector.IDs)){
  sector.summary<-subset(match.summary, Use.Type==sector.IDs[s])
  an.values<-quantile(sector.summary$CU_scale, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  warm.values<-quantile(sector.summary$CU_scale_warm, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  cool.values<-quantile(sector.summary$CU_scale_cool, probs=c(0.25,0.5,0.75), na.rm=TRUE)
  cols<-c((s*3-2):(s*3))
  quant.summary[1,cols] <- an.values 
  quant.summary[2,cols] <- warm.values 
  quant.summary[3,cols] <- cool.values 
}

setwd("G:/My Drive/VT/Research/Virginia/USGS_WUDR_ConsumptiveUse/Analysis/FacilityAnalysis")
write.csv(xSecSummary, "xSecSummary_facilities.csv", row.names=TRUE)
write.csv(XSecRevised, "xSecSummary_facilities_ConfInt.csv", row.names=TRUE)
write.csv(quant.summary, "QuantSummary_facilities.csv", row.names=TRUE)


save.image("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis\\VA_FacilityAnalysis.RData")
#load("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\FacilityAnalysis\\VA_FacilityAnalysis.RData")



prob.records<-subset(match.summary,CU_full < -1) # 12 facilities with CU less than -1
clean.records<-subset(match.summary, CU_full > -1) # 90 facilities CU greater than -1 (NA values removed)
summary(prob.records$VPDES.Nrecord)   # mean of 11 withdrawal records, 15 discharge
summary(clean.records$VPDES.Nrecord)  # mean of 31 withdrawal records, 31 discharge

summary(clean.records)
hist(clean.records$VWUDS.Nrecord)

hist(clean.records$CU_full[clean.records$Use.Type=="Energy"], main="Energy")
hist(clean.records$CU_full[clean.records$Use.Type=="Industrial"], main="Industrial")
hist(clean.records$CU_full[clean.records$Use.Type=="Commercial"], main="Commercial")

write.csv(clean.records,"CleanRecords.csv",row.names=FALSE)
