==> JAWRA_Files/06_VA_Statewide_Consumptive_Use.R <==
###############################################################################################################################################
#########################################################State Level Analysis##################################################################

#Primary Author: Morgan McCarthy, M.S. Biological Systems Engineering, Virginia Tech

#-----------------------------------------------------------Purpose---------------------------------------------------------------------------#

# The purpose of this script is to compare withdrawals, discharges, and consumption over time on a state-wide level for the Commonwealth of
# Virginia. Discharge data is withdrawn from the EPA's ECHO system, using the ECHO_Timeseries.R script. This data is then cleaned in the 
# ECHO_QAQC.R script. Withdrawal data is obtained from the VDEQ's VA Hydro system, which is only available through collaboration. Please 
# contact the Office of Water Supply if you are interested in collaboration. Withdrawal data is cleaned in the VWUDS_QAQC.R script. Water
# sectors for each facility in both datasets are classified in the ECHO_QAQC.R and VWUDS_QAQC.R scripts. 

# Consumption, in this analysis, is defined as water that is permanently removed from a water resource system and not replaced. It is 
# calculated as a water balance, with water in being discharges and water out being withdrawals. 

##################################################Library Initilization########################################################################
rm(list=ls())

#library(foreign) #allows for easy manipulation of *.dbf file types
#library(rgdal) #extract files from file geodatabases-like in ArcMap
library(dplyr)#data manipulation package that speeds up grouping, summarizing, ordering. 
#library(XML) #downloads and reads XML file types-used to access ECHORest Services to download facilities
#library(RCurl) #downloads and interacts with web-based content
#library(readxl) #reads excel files
#library(jsonlite)
#library(httr)
#library(proj4)
#library(data.table)
#library(stringr)
#library(plotly)
#library(shiny)
#library(lubridate)
#library(RColorBrewer)
library(tibble)
#library(tidyr)
#library(Hmisc)
#library(gridExtra)
library(ggplot2)

# For using plotly 
#Sys.setenv("plotly_username" = "mccartma")
#Sys.setenv("plotly_api_key" = "r3SDikk9PJVh3dxt6F5q") # update api key each new day https://plot.ly/settings/api#/

options(scipen=999) #Disables scientific notation

################################################################Inputs########################################################################

#----------------------------------------------------Cleaned Discharge Data-----------------------------------------------------------------#
# This discharge data was generated by the ECHO_Timeseries.R script and cleaned in the ECHO_QAQC.R Script

#--Load Discharge Timeseries Data--#
setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\StatewideAnalysis")
ECHO_2010_2017<-read.table("ECHO_2010_2017_QAQC.txt", sep="\t", header=T)
## Unique facilities in discharge data
Fac.list<-unique(ECHO_2010_2017$VPDES.Name)

#--Facilities with Matched Withdrawal Facility--#
Matched<-read.csv("Runninglist_Matches.csv")
Matched<-mutate_if(Matched,is.factor,as.character)

#--Remove Municipal Data because it has a noisy relationship--#
Matched<-subset(Matched,!Matched$Sector=="Municipal")

# This function looks at the list of matched facilities and pulls relevant attributes and DMR data for the discharging facilities
## Just run, not as function
#matched_discharge<- function(Matched){
  
Match_ECHO_2010_2017<-subset(ECHO_2010_2017,ECHO_2010_2017$Facility.ID%in%gsub("echo_","",Matched$VPDES.Hydrocode)
  , select=c(1,3:9,12,19,33,43,44,45,46))

colnames(Match_ECHO_2010_2017)<-c("OutfallID","VPDES.Facility.ID","VPDES.Name","Fac.Lat","Fac.Long","Outfall_Type","VPDES.Outfall.Long",
                                  "VPDES.Outfall.Lat", "Discharges_MGD", "Date","County","Waterbody","Use.Type","Fuel_Type","Gen_Cap_MW")
Match_ECHO_2010_2017$MatchID<-Matched$MatchID[match(Match_ECHO_2010_2017$VPDES.Facility.ID,gsub("echo_","",Matched$VPDES.Hydrocode))]
Match_ECHO_2010_2017$Date<-as.Date(Match_ECHO_2010_2017$Date,format="%Y-%m-%d")

ECHO_2010_2017<-subset(ECHO_2010_2017,select=c(1:9,10,12,19,33,40,43,44,45,46,47))
colnames(ECHO_2010_2017)<-c("OutfallID","Year","VPDES.Facility.ID","VPDES.Name","Fac.Lat","Fac.Long","Outfall_Type",
                            "Outfall.Long","Outfall.Lat","Measured_Effluent","Discharges_MGD","Date","County",
                            "Permit_Type","Waterbody","Use.Type","Fuel_Type","Gen_Cap_MW","Mon_Reported")
ECHO_2010_2017$Date<-as.Date(ECHO_2010_2017$Date,format="%Y-%m-%d")

assign("ECHO_2010_2017",ECHO_2010_2017,envir=.GlobalEnv)
assign("Match_ECHO_2010_2017",Match_ECHO_2010_2017,envir=.GlobalEnv)
#write.table(ECHO_2010_2017,file="G:/My Drive/ECHO NPDES/USGS_Consumptive_Use_Updated/ECHO_2010_2017_QAQC_statewide.txt",sep="\t")

#}
#matched_discharge(Matched)

#------------------------------------------------------------------------------------------------------------------------------------------#
#-----------------------------------------------Cleaned Withdrawal Data--------------------------------------------------------------------#
# Withdrawal data was obtained through VDEQ's VA Hydro site, which can only be obtained through collaboration.
# Data was analyzed, coordinates were corrected, and water sectors were classified in VWUDS_QAQC.R Script. 

#--Load Withdrawal Timeseries Data--#
load("G:/My Drive/VT/Research/Virginia/USGS_WUDR_ConsumptiveUse/Analysis/StatewideAnalysis/VWUDS_2010_2017.RData")
#VWUDS_2010_2017<-VWUDS_2010_2017[VWUDS_2010_2017$Date %within% interval("2010-01-01","2016-12-31"),]

# This function looks at the list of matched facilities and pulls relevant attributes and DMR data for the withdrawing facilities
#matched_withdrawal<- function(Matched){
# Just run code, not as function
#VWUDS_2010_2017<-as.data.table(VWUDS_2010_2017)
VWUDS_2010_2017<-VWUDS_2010_2017[order(VWUDS_2010_2017[,1],VWUDS_2010_2017[,8],decreasing=F),]
colnames(VWUDS_2010_2017)[3]<-"VWUDS.Facility.ID"
colnames(VWUDS_2010_2017)[10]<-"Old.Use.Type"
colnames(VWUDS_2010_2017)[20]<-"Use.Type"

#--Facilities with Matched Discharge Facility--#
Match_VWUDS_2010_2017<-subset(VWUDS_2010_2017,VWUDS_2010_2017$VWUDS.Facility.ID%in%Matched$VWUDS.HydroID)
Match_VWUDS_2010_2017<-Match_VWUDS_2010_2017%>%
  dplyr::group_by(VWUDS.Facility.ID,Date)%>%
  dplyr::summarise(VWUDS.Name=first(Facility),Withdrawal_mgd=sum(Withdrawals_MGD,na.rm=T),VWUDS.Lat=mean(Corrected_Latitude,na.rm=T),VWUDS.Long=mean(Corrected_Longitude,na.rm=T),
                   VWUDS.Use.Type=first(Use.Type))

colnames(Match_VWUDS_2010_2017)<-c("VWUDS.Facility.ID","Date","VWUDS.Name","Withdrawals_MGD","VWUDS.Lat",
                                   "VWUDS.Long","VWUDS.Use.Type")
Match_VWUDS_2010_2017$MatchID<-Matched$MatchID[match(Match_VWUDS_2010_2017$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]

assign("VWUDS_2010_2017",VWUDS_2010_2017,envir=.GlobalEnv)
assign("Match_VWUDS_2010_2017",Match_VWUDS_2010_2017,envir=.GlobalEnv)

save(VWUDS_2010_2017,file="G:/My Drive/ECHO NPDES/USGS_Consumptive_Use_Updated/Code/R Workspaces/VWUDS_2010_2017_statewide.RData")

#}
#matched_withdrawal(Matched)

#-----------------------------Matched Data---------------------------------#
# Look for "fully" matched results for each month. This function merges DMR entries reported by both the withdrawing and discharging 
# facilities. This ensures that we aren't looking at incomplete data. i.e. facilities that report a discharge but not a withdrawal for that reporting period.

# We are also concerned with energy facilities that report a 0.0 MGD discharge or withdrawal. Is it becuase they didn't report, or because it was actually 0?
# Just run code, not function
#fully_matched<- function(Match_VWUDS_2010_2017,Match_ECHO_2010_2017){
full_match_2010_2017<-merge(Match_VWUDS_2010_2017,Match_ECHO_2010_2017,by=c("MatchID","Date"))
full_match_2010_2017<-full_match_2010_2017[,c("VPDES.Facility.ID","OutfallID","VWUDS.Facility.ID","VWUDS.Name","VPDES.Name","Use.Type",
                      "Fac.Lat","Fac.Long","VWUDS.Lat","VWUDS.Long","Date","Withdrawals_MGD","Discharges_MGD","County","Waterbody","Fuel_Type",
                      "Gen_Cap_MW")]


#---------------------Dissolve Flow from multiple outfalls into a single facility value----------#
full_match_2010_2017<-full_match_2010_2017%>%dplyr::group_by(VWUDS.Facility.ID,VPDES.Facility.ID,
                                       VWUDS.Name,VPDES.Name,Use.Type,Fac.Lat,Fac.Long,
                                       VWUDS.Lat,VWUDS.Long,Date,Withdrawals_MGD,
                                       County,Waterbody,Fuel_Type,Gen_Cap_MW)%>%dplyr::summarise(Discharges_MGD=sum(Discharges_MGD,na.rm=T),n_outfalls=n_distinct(OutfallID))
full_match_2010_2017<-full_match_2010_2017[c("VWUDS.Facility.ID","VPDES.Facility.ID","n_outfalls",
                                             "VWUDS.Name","VPDES.Name","Use.Type","Fac.Lat","Fac.Long",
                                             "VWUDS.Lat","VWUDS.Long","Date","Withdrawals_MGD", "Discharges_MGD",
                                             "County","Waterbody","Fuel_Type",
                                             "Gen_Cap_MW")]

full_match_2010_2017<-full_match_2010_2017%>%add_column(CU=(full_match_2010_2017$Withdrawals_MGD-full_match_2010_2017$Discharges_MGD)/full_match_2010_2017$Withdrawals_MGD, .after="Discharges_MGD")


full_match_2010_2017_summary<-full_match_2010_2017%>%dplyr::group_by(VPDES.Facility.ID)%>%
  dplyr::summarise(VWUDS.Facility.ID=first(VWUDS.Facility.ID),n_outfalls=first(n_outfalls),VWUDS.Name=first(VWUDS.Name),VPDES.Name=first(VPDES.Name),
                   Use.Type=first(Use.Type), VPDES.Fac.Lat=first(Fac.Lat),VPDES.Fac.Long=first(Fac.Long),VWUDS.Fac.Lat=first(VWUDS.Lat),VWUDS.Fac.Long=first(VWUDS.Long),
                   Ave_Withdrawal_mgd=mean(Withdrawals_MGD,na.rm=T),Ave_Discharge_mgd=mean(Discharges_MGD,na.rm=T),County=first(County),Waterbody=first(Waterbody),
                   Fuel_Type=first(Fuel_Type),Gen_Cap_MW=first(Gen_Cap_MW))

full_match_2010_2017_summary$Ave_CU<-(full_match_2010_2017_summary$Ave_Withdrawal_mgd-full_match_2010_2017_summary$Ave_Discharge_mgd)/full_match_2010_2017_summary$Ave_Withdrawal_mgd

full_match_2010_2017_summary$VPDES_perc<-Matched$VPDES...total[match(full_match_2010_2017_summary$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]
full_match_2010_2017_summary$VWUDS_perc<-Matched$VWUDS...total[match(full_match_2010_2017_summary$VWUDS.Facility.ID,Matched$VWUDS.HydroID)]

                                 
# Not all facilities report 12 months out of the year. Therefore, if we want to aggregate monthly withdrawals into yearly estimates for each
# facility, we must divide by the number of reporting months.
W_Months<-
  full_match_2010_2017%>%
  dplyr::group_by(VWUDS.Facility.ID,Year=substring(Date,1,4))%>%
  dplyr::count(Month=substring(Date,6,7))%>%
  dplyr::summarise(Mon_Reported=sum(n))

full_match_2010_2017<-full_match_2010_2017%>%add_column(Year=substring(full_match_2010_2017$Date,1,4), .after="Date")
full_match_2010_2017<-merge(full_match_2010_2017,W_Months,by=c("VWUDS.Facility.ID","Year"),all.x=T)


#############################################################Analysis########################################################################

#-------------------------------------------------------Timeseries of Withdrawals vs Discharges-------------------------------------------#
## NEW CODE FROM ME:
load("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\StatewideAnalysis\\StatewideCU.RData")
VWUDS.uses<-as.factor(VWUDS_2010_2017$Use.Type)

## All Nonenergy
################################
VWUDS.nonenergy <- subset(VWUDS_2010_2017,Use.Type != "Energy")
wd.nonenergy<-VWUDS.nonenergy%>%
  dplyr::group_by(Date)%>% # you don't have to divide by 12 months here because the facility dataframe already does that
  dplyr::summarise(Withdrawals_MGD=sum(Withdrawals_MGD),
                   nFacilities=n_distinct(VWUDS.Facility.ID))
plot(wd.nonenergy$Date, wd.nonenergy$Withdrawals_MGD, type ="l")

ECHO.nonenergy <- subset(ECHO_2010_2017,Use.Type != "Energy")
dc.nonenergy<-ECHO.nonenergy%>%
  dplyr::group_by(Date)%>% 
  dplyr::summarise(Discharges_MGD=sum(Discharges_MGD, na.rm=TRUE),
                   nFacilities=n_distinct(VPDES.Facility.ID))
plot(dc.nonenergy$Date, dc.nonenergy$Discharges_MGD, type ="l")

# trim post-2016 wd data
wd.nonenergy<-wd.nonenergy[ wd.nonenergy$Date %in% dc.nonenergy$Date ,]
summary(dc.nonenergy$Date)
summary(wd.nonenergy$Date)

## CONVERT WD AND DC TO MCM PER DAY
wd.nonenergy$Withdrawals_MCM<-wd.nonenergy$Withdrawals_MGD / 264.172 
dc.nonenergy$Discharges_MCM<-dc.nonenergy$Discharges_MGD / 264.172 

# calculate consumption and scale it so it can be plotted alongside wd and dc
cons.nonenergy<- (wd.nonenergy$Withdrawals_MGD - dc.nonenergy$Discharges_MGD) / wd.nonenergy$Withdrawals_MGD
cons.scale.nonenergy<-cons.nonenergy * max(wd.nonenergy$Withdrawals_MCM)
axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.nonenergy$Withdrawals_MCM)

par(mar=c(5.1, 4.1, 4.1, 3.5), mgp= c(2,0.8,0))
plot(wd.nonenergy$Date, wd.nonenergy$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(wd.nonenergy$Withdrawals_MCM)), main = "All Nonenergy Facilities", 
     xlab="Date", ylab="Volume (MCM per day)" )
  lines(dc.nonenergy$Date, dc.nonenergy$Discharges_MCM, type ="l", col= "blue")
  lines(wd.nonenergy$Date, cons.scale.nonenergy, lty=3, lwd=1.5)
  axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))
  mtext("Consumption", side=4, line=2)
  
## Industrial
################################
VWUDS.industrial <- subset(VWUDS_2010_2017,Use.Type == "Industrial")
wd.industrial<-VWUDS.industrial%>%
    dplyr::group_by(Date)%>% # you don't have to divide by 12 months here because the facility dataframe already does that
    dplyr::summarise(Withdrawals_MGD=sum(Withdrawals_MGD),
                     nFacilities=n_distinct(VWUDS.Facility.ID))
  plot(wd.industrial$Date, wd.industrial$Withdrawals_MGD, type ="l")
  
ECHO.industrial <- subset(ECHO_2010_2017,Use.Type == "Industrial")
dc.industrial<-ECHO.industrial%>%
    dplyr::group_by(Date)%>% 
    dplyr::summarise(Discharges_MGD=sum(Discharges_MGD, na.rm=TRUE),
                     nFacilities=n_distinct(VPDES.Facility.ID))
  plot(dc.industrial$Date, dc.industrial$Discharges_MGD, type ="l")
  
# trim post-2016 wd data
wd.industrial<-wd.industrial[ wd.industrial$Date %in% dc.industrial$Date ,]
  summary(dc.industrial$Date)
  summary(wd.industrial$Date)
  
## CONVERT WD AND DC TO MCM PER DAY
wd.industrial$Withdrawals_MCM<-wd.industrial$Withdrawals_MGD / 264.172 
dc.industrial$Discharges_MCM<-dc.industrial$Discharges_MGD / 264.172 
  
# calculate consumption and scale it so it can be plotted alongside wd and dc
cons.industrial<- (wd.industrial$Withdrawals_MGD - dc.industrial$Discharges_MGD) / wd.industrial$Withdrawals_MGD
cons.scale.ind<-cons.industrial * max(wd.industrial$Withdrawals_MCM)
axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.industrial$Withdrawals_MCM)
  
par(mar=c(5.1, 4.1, 4.1, 3.5), mgp= c(2,0.8,0))
plot(wd.industrial$Date, wd.industrial$Withdrawals_MCM, type ="l", col= "orange", 
       ylim=c(0,max(wd.industrial$Withdrawals_MCM)), main = "Industrial Facilities", 
       xlab="Date", ylab="Volume (MCM per day)")
  lines(dc.industrial$Date, dc.industrial$Discharges_MCM, type ="l", col= "blue")
  lines(wd.industrial$Date, cons.scale.ind, lty=3, lwd=1.5)
  axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))
  
## Municipal
################################
VWUDS.municipal <- subset(VWUDS_2010_2017,Use.Type == "Municipal")
wd.municipal<-VWUDS.municipal%>%
  dplyr::group_by(Date)%>% # you don't have to divide by 12 months here because the facility dataframe already does that
  dplyr::summarise(Withdrawals_MGD=sum(Withdrawals_MGD),
                     nFacilities=n_distinct(VWUDS.Facility.ID))
plot(wd.municipal$Date, wd.municipal$Withdrawals_MGD, type ="l")
  
ECHO.municipal <- subset(ECHO_2010_2017,Use.Type == "Municipal")
dc.municipal<-ECHO.municipal%>%
  dplyr::group_by(Date)%>% 
  dplyr::summarise(Discharges_MGD=sum(Discharges_MGD, na.rm=TRUE),
                     nFacilities=n_distinct(VPDES.Facility.ID))
  plot(dc.municipal$Date, dc.municipal$Discharges_MGD, type ="l")
  
# trim post-2016 wd data
wd.municipal<-wd.municipal[ wd.municipal$Date %in% dc.municipal$Date ,]
  summary(dc.municipal$Date)
  summary(wd.municipal$Date)
  
## CONVERT WD AND DC TO MCM PER DAY
wd.municipal$Withdrawals_MCM<-wd.municipal$Withdrawals_MGD / 264.172 
dc.municipal$Discharges_MCM<-dc.municipal$Discharges_MGD / 264.172 
  
# calculate consumption and scale it so it can be plotted alongside wd and dc
cons.municipal<- (wd.municipal$Withdrawals_MGD - dc.municipal$Discharges_MGD) / wd.municipal$Withdrawals_MGD
cons.scale.mun<-cons.municipal * max(wd.municipal$Withdrawals_MCM)
axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.municipal$Withdrawals_MCM)
  
par(mar=c(5.1, 4.1, 4.1, 3.5), mgp= c(2,0.8,0))
plot(wd.municipal$Date, wd.municipal$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(wd.municipal$Withdrawals_MCM)), main = "Municipal Facilities", 
     xlab="Date", ylab="Volume (MCM per day)")
lines(dc.municipal$Date, dc.municipal$Discharges_MCM, type ="l", col= "blue")
lines(wd.municipal$Date, cons.scale.mun, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))  
  
## Save plot to file
#################################

setwd("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\StatewideAnalysis")
ppi<-300

png(file="CU_TS_allFacilities_R1.png",width=6*ppi,height=6*ppi,res=ppi)		
par(mfrow=c(3,1))
par(mar=c(2, 3, 2, 3.5), mgp= c(2,0.8,0))

axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.nonenergy$Withdrawals_MCM)
date.axis<-as.Date(c("2011-01-01","2013-01-01","2015-01-01"))

plot(wd.nonenergy$Date, wd.nonenergy$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(wd.nonenergy$Withdrawals_MCM)), main = "",
     xlab="", ylab="Volume (MCM per day)" )
title("All Nonenergy Facilities", adj = 0.5, line = 0.6)
lines(dc.nonenergy$Date, dc.nonenergy$Discharges_MCM, type ="l", col= "blue")
lines(wd.nonenergy$Date, cons.scale.nonenergy, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))
axis (side = 1, at = date.axis, labels = FALSE)
mtext("Consumptive Use", side=4, line=2, cex=0.7)
legend("bottomleft",legend=c("Withdrawals","Discharges","Consumptive Use"),
       col=c("orange","blue","black"), lty=c(1,1,3), ncol=3, bty="n")

axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.industrial$Withdrawals_MCM)
plot(wd.industrial$Date, wd.industrial$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(wd.industrial$Withdrawals_MCM)), main = "", 
     xlab="", ylab="Volume (MCM per day)")
title("Industrial Facilities", adj = 0.5, line = 0.6)
lines(dc.industrial$Date, dc.industrial$Discharges_MCM, type ="l", col= "blue")
lines(wd.industrial$Date, cons.scale.ind, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))
axis (side = 1, at = date.axis, labels = FALSE)
mtext("Consumptive Use", side=4, line=2, cex=0.7)

axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(wd.municipal$Withdrawals_MCM)
plot(wd.municipal$Date, wd.municipal$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(wd.municipal$Withdrawals_MCM)), main = "", 
     xlab="", ylab="Volume (MCM per day)")
title("Municipal Facilities", adj = 0.5, line = 0.6)
lines(dc.municipal$Date, dc.municipal$Discharges_MCM, type ="l", col= "blue")
lines(wd.municipal$Date, cons.scale.mun, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0)) 
axis (side = 1, at = date.axis, labels = FALSE)
mtext("Consumptive Use", side=4, line=2, cex=0.7)

dev.off()

#####################################################
## MATCHED FACILITIES
#####################################################  

energy.match <- subset(full_match_2010_2017,Use.Type == "Energy")
statewide.energy.match<-energy.match%>%
  dplyr::group_by(Date)%>% 
  dplyr::summarise(Withdrawals_MGD=sum(Withdrawals_MGD),
                   Discharges_MGD=sum(Discharges_MGD))

## CONVERT WD AND DC TO MCM PER DAY
statewide.energy.match$Withdrawals_MCM<-statewide.energy.match$Withdrawals_MGD / 264.172 
statewide.energy.match$Discharges_MCM<-statewide.energy.match$Discharges_MGD / 264.172 

# calculate consumption and scale it so it can be plotted alongside wd and dc
cons.match.energy<- (statewide.energy.match$Withdrawals_MGD - statewide.energy.match$Discharges_MGD) / statewide.energy.match$Withdrawals_MGD
cons.scale.match.energy<-cons.match.energy * max(statewide.energy.match$Withdrawals_MCM)


## Industrial
industrial.match <- subset(full_match_2010_2017,Use.Type == "Industrial")
statewide.ind.match<-industrial.match%>%
  dplyr::group_by(Date)%>% 
  dplyr::summarise(Withdrawals_MGD=sum(Withdrawals_MGD),
                   Discharges_MGD=sum(Discharges_MGD))

## CONVERT WD AND DC TO MCM PER DAY
statewide.ind.match$Withdrawals_MCM<-statewide.ind.match$Withdrawals_MGD / 264.172 
statewide.ind.match$Discharges_MCM<-statewide.ind.match$Discharges_MGD / 264.172 

# calculate consumption and scale it so it can be plotted alongside wd and dc
cons.match.ind<- (statewide.ind.match$Withdrawals_MGD - statewide.ind.match$Discharges_MGD) / statewide.ind.match$Withdrawals_MGD
cons.scale.match.ind<-cons.match.ind * max(statewide.ind.match$Withdrawals_MCM)

## Save plot to file
png(file="CU_TS_MatchFacilities_R1.png",width=6*ppi,height=5*ppi,res=ppi)		
par(mfrow=c(2,1))
par(mar=c(2, 3, 2, 3.5), mgp= c(2,0.8,0))

axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(statewide.energy.match$Withdrawals_MCM)
plot(statewide.energy.match$Date, statewide.energy.match$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(statewide.energy.match$Withdrawals_MCM)), main = "", 
     xlab="", ylab="Volume (MCM per day)", cex.axis=0.8, cex.lab=0.8)
title("Matched Energy Facilities", adj = 0.5, line = 0.6, cex.main=0.8)
lines(statewide.energy.match$Date, statewide.energy.match$Discharges_MCM, type ="l", col= "blue")
lines(statewide.energy.match$Date, cons.scale.match.energy, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0), cex.axis=0.8, cex.lab=0.8) 
axis (side = 1, at = date.axis, labels = FALSE)
mtext("Consumptive Use", side=4, line=2, cex=0.8)


axis.vals<-c(0,0.2,0.4,0.6,0.8,1.0) * max(statewide.ind.match$Withdrawals_MCM)
plot(statewide.ind.match$Date, statewide.ind.match$Withdrawals_MCM, type ="l", col= "orange", 
     ylim=c(0,max(statewide.ind.match$Withdrawals_MCM)), main = "", 
     xlab="", ylab="Volume (MCM per day)", cex.axis=0.8, cex.lab=0.8)
title("Matched Industrial Facilities", adj = 0.5, line = 0.6, cex.main=0.8)
lines(statewide.ind.match$Date, statewide.ind.match$Discharges_MCM, type ="l", col= "blue")
lines(statewide.ind.match$Date, cons.scale.match.ind, lty=3, lwd=1.5)
axis (side = 4, at = axis.vals, labels = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0), cex.axis=0.8, cex.lab=0.8) 
axis (side = 1, at = date.axis, labels = FALSE)
mtext("Consumptive Use", side=4, line=2, cex=0.8)
legend("bottomright",legend=c("Withdrawals","Discharges","Consumptive Use"),
       col=c("orange","blue","black"), lty=c(1,1,3), ncol=3, bty="n", cex=0.8)

dev.off()


save.image("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\StatewideAnalysis\\StatewideCU.RData")
load("G:\\My Drive\\VT\\Research\\Virginia\\USGS_WUDR_ConsumptiveUse\\Analysis\\StatewideAnalysis\\StatewideCU.RData")
